Update project files and fix wiki page retrieval
Updates project files in this repo to use more modern formats
usable by Visual Studio Community 2017 and later. Note that this
is just in this base repo. Submodules will need to be updated and
changed here later.

In addition, this fixes wiki page retrieval due to changes in the
<text> element from the Export. It no longer immediately has the
xml attribute afterwards.